+++ lrg_omics/proteomics/maxquant/pipeline.py
@@ -5,7 +5,7 @@ import lrg_omics
 from os.path import isfile, basename, join, abspath, dirname
 from datetime import date
 from pathlib import Path as P
-from ..common import maybe_make_dir_and_chdir, maybe_create_symlink
+from ...common import maybe_make_dir_and_chdir, maybe_create_symlink
 
 FAKEPATH = abspath("tests/data/maxquant/tmt11/txt")
