--- lrg_omics/proteomics/pipelines/common.py
@@ -1,6 +1,0 @@
-import os
-
-
-def maybe_create_symlink(src, dst):
-    if not os.path.isfile(dst):
-        os.symlink(src, dst)
