--- lrg_omics/tools.py
@@ -1,5 +1,0 @@
-from datetime import date
-
-
-def today():
-    return date.today().strftime("%y%m%d")
